+++ src/main/java/legend/game/types/WmapSmokeCloudInstance60.java
@@ -5,6 +5,6 @@ import legend.core.gte.SVECTOR;
 
 public class WmapSmokeCloudInstance60 {
   public final GsCOORDINATE2 coord2_00 = new GsCOORDINATE2();
-  public int scaleAndColourFade_50;
+  public float scaleAndColourFade_50;
   public final SVECTOR translationOffset_54 = new SVECTOR();
 }
